fix: AvgTokensをfloat64型に修正
SQLのAVG()関数はfloat64を返すため、AvgTokensフィールドを
int型からfloat64型に変更。

修正箇所:
- ProjectStats.AvgTokens
- GroupStats.AvgTokens
- ProjectStatsResponse.AvgTokens
- ProjectGroupStatsResponse.AvgTokens

+++ internal/api/types.go
@@ -125,7 +125,7 @@ type ProjectStatsResponse struct {
 	TotalCacheCreationTokens int       `json:"totalCacheCreationTokens"`
 	TotalCacheReadTokens     int       `json:"totalCacheReadTokens"`
 	TotalTokens              int       `json:"totalTokens"`
-	AvgTokens                int       `json:"avgTokens"`
+	AvgTokens                float64   `json:"avgTokens"`
 	FirstSession             time.Time `json:"firstSession"`
 	LastSession              time.Time `json:"lastSession"`
 	ErrorRate                float64   `json:"errorRate"`
@@ -193,7 +193,7 @@ type ProjectGroupStatsResponse struct {
 	TotalOutputTokens        int       `json:"totalOutputTokens"`
 	TotalCacheCreationTokens int       `json:"totalCacheCreationTokens"`
 	TotalCacheReadTokens     int       `json:"totalCacheReadTokens"`
-	AvgTokens                int       `json:"avgTokens"`
+	AvgTokens                float64   `json:"avgTokens"`
 	FirstSession             time.Time `json:"firstSession"`
 	LastSession              time.Time `json:"lastSession"`
 	ErrorRate                float64   `json:"errorRate"`
